+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 - Replacements use either `$10` or `${1}0`
 - Specification update -> Version 1.1 Final
 - Travis Build added
+- Quick tests (quick-tests.json) for npm package added.
+  The big test file gets excluded.